fix: make winners pop-up as a separate component
